+++ libpod/finished_unsupported.go
@@ -1,0 +1,12 @@
+// +build darwin,!linux
+
+package libpod
+
+import (
+	"os"
+	"time"
+)
+
+func getFinishedTime(fi os.FileInfo) time.Time {
+	return time.Time{}
+}
