+++ libpod/oci_unsupported.go
@@ -2,6 +2,19 @@
 
 package libpod
 
-func moveConmonToCgroup() error {
+import (
+	"os"
+	"os/exec"
+)
+
+func (r *OCIRuntime) moveConmonToCgroup(ctr *Container, cgroupParent string, cmd *exec.Cmd) error {
 	return ErrOSNotSupported
+}
+
+func newPipe() (parent *os.File, child *os.File, err error) {
+	return nil, nil, ErrNotImplemented
+}
+
+func (r *OCIRuntime) createContainer(ctr *Container, cgroupParent string) (err error) {
+	return ErrNotImplemented
 }
